Refactor: split monolithic main.go into internal packages
Split 995-line main.go and 2154-line main_test.go into 6 internal
packages (config, socks, metrics, tunnel, checker, testutil) with
14 improvements: slog logging, custom Prometheus registry, io.ReadFull
in SOCKS dialer, reusable http.Client per tunnel, config validation
before reload, configurable socks_base_port, TunnelInitErrors metric,
%w error wrapping, /health endpoint, HTTP server timeouts, and
golangci-lint v2 CI integration.
